requirements.txt
flask
requests
pyyaml
pyjwt[crypto]
Flask-Session